requirements.txt
anthropic
flask
chromadb
pdfplumber
beautifulsoup4
lxml
sentence-transformers
python-dotenv
